backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
anthropic
python-multipart
pypdf
